refactor(localdebug): move localdebug into solution
